SmartStreamer: Add user settings overlaying support.
As a proof-of-concept, I add user overlay support for encoders.json.
With this change, all changes encoders.json will be redirected to a another
overlay file, which will be the user settings.

When we read settings, we will overlay settings from overlay file onto
original settings and all classes in the application will use these settings.

We don't overlay all settings blindly, we're keeping a list of keys and
only permitting those settings to be overlayed. For instance, encoders.json
has these keys defined:

	encoderOverlayKeys << "bitrate";
	encoderOverlayKeys << "frameRate";
	encoderOverlayKeys << "shaping";
	encoderOverlayKeys << "resolution";
	encoderOverlayKeys << "multicast_address_base";

If this works as expected, other settings will be implemented in later commits.